move: Auction 응답, 요청 DTO 파일 분리 (#CHZZ-35)
