+++ src/main/java/org/chzz/market/domain/auction/dto/response/AuctionResponse.java
@@ -1,10 +1,9 @@
-package org.chzz.market.domain.auction.dto;
+package org.chzz.market.domain.auction.dto.response;
 
 import com.querydsl.core.annotations.QueryProjection;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import lombok.Getter;
-import org.chzz.market.domain.auction.entity.Auction;
 
 /**
  * 진행중인 경매 목록 조회 DTO
