+++ src/main/java/org/chzz/market/domain/auction/repository/AuctionRepositoryCustom.java
@@ -1,8 +1,8 @@
 package org.chzz.market.domain.auction.repository;
 
 import java.util.Optional;
-import org.chzz.market.domain.auction.dto.AuctionDetailsResponse;
-import org.chzz.market.domain.auction.dto.AuctionResponse;
+import org.chzz.market.domain.auction.dto.response.AuctionDetailsResponse;
+import org.chzz.market.domain.auction.dto.response.AuctionResponse;
 import org.chzz.market.domain.product.entity.Product.Category;
 import org.chzz.market.domain.auction.entity.SortType;
 import org.springframework.data.domain.Page;
