+++ src/test/java/org/chzz/market/domain/auction/repository/AuctionRepositoryImplTest.java
@@ -6,9 +6,10 @@ import static org.chzz.market.domain.product.entity.Product.*;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
+
 import org.chzz.market.common.DatabaseTest;
-import org.chzz.market.domain.auction.dto.AuctionDetailsResponse;
-import org.chzz.market.domain.auction.dto.AuctionResponse;
+import org.chzz.market.domain.auction.dto.response.AuctionDetailsResponse;
+import org.chzz.market.domain.auction.dto.response.AuctionResponse;
 import org.chzz.market.domain.auction.entity.Auction;
 import org.chzz.market.domain.auction.entity.Auction.AuctionStatus;
 import org.chzz.market.domain.auction.entity.SortType;
@@ -68,10 +69,10 @@ class AuctionRepositoryImplTest {
         user3 = User.builder().providerId("123456").nickname("닉네임3").email("asd12@naver.com").build();
         userRepository.saveAll(List.of(user1, user2, user3));
 
-        product1 = builder().user(user1).name("제품1").category(Category.FASHION_AND_CLOTHING).build();
-        product2 = builder().user(user1).name("제품2").category(Category.BOOKS_AND_MEDIA).build();
-        product3 = builder().user(user2).name("제품3").category(Category.FASHION_AND_CLOTHING).build();
-        product4 = builder().user(user2).name("제품4").category(Category.FASHION_AND_CLOTHING).build();
+        product1 = builder().user(user1).name("제품1").category(Category.FASHION_AND_CLOTHING).minPrice(10000).build();
+        product2 = builder().user(user1).name("제품2").category(Category.BOOKS_AND_MEDIA).minPrice(20000).build();
+        product3 = builder().user(user2).name("제품3").category(Category.FASHION_AND_CLOTHING).minPrice(30000).build();
+        product4 = builder().user(user2).name("제품4").category(Category.FASHION_AND_CLOTHING).minPrice(40000).build();
         productRepository.saveAll(List.of(product1, product2, product3, product4));
 
         auction1 = Auction.builder().product(product1).minPrice(1000).status(AuctionStatus.PROCEEDING)
