+++ src/test/java/org/chzz/market/util/AuctionTestFactory.java
@@ -1,6 +1,6 @@
 package org.chzz.market.util;
 
-import org.chzz.market.domain.auction.dto.RegisterAuctionRequest;
+import org.chzz.market.domain.auction.dto.request.RegisterAuctionRequest;
 import org.chzz.market.domain.auction.entity.Auction;
 import org.chzz.market.domain.product.entity.Product;
 import org.springframework.test.util.ReflectionTestUtils;
